Updated remove member to unassign member as well.

+++ src/app/api/projects/[id]/members/[memberId]/route.ts
@@ -13,7 +13,36 @@ export async function DELETE(request: Request, { params: { id, memberId } }: Mem
         return unauthorizedResponse({ success: false, data: "Unauthorized" });
     }
 
-    const { error } = await getServiceSupabase()
+    const serviceSupabase = getServiceSupabase();
+
+    const { data: assigneeData, error: assigneeError } = await serviceSupabase
+        .from("task_assignee")
+        .select("task!inner(project_id, task_id)")
+        .eq("task.project_id", id)
+        .eq("user_id", memberId);
+
+    if (assigneeError) {
+        console.error(assigneeError);
+        return internalErrorResponse({ success: false, data: "Unable to remove that member" });
+    }
+
+    const { error: deleteError } = await serviceSupabase
+        .from("task_assignee")
+        .delete()
+        .in(
+            "task_id",
+            (assigneeData ?? []).map((assignee) => assignee.task!.task_id),
+        );
+
+    if (deleteError) {
+        console.error(deleteError);
+        return internalErrorResponse({
+            success: false,
+            data: "Unable to remove that member. Please try again.",
+        });
+    }
+
+    const { error } = await serviceSupabase
         .from("project_member")
         .delete()
         .eq("user_id", memberId)
@@ -21,7 +50,7 @@ export async function DELETE(request: Request, { params: { id, memberId } }: Mem
 
     if (error) {
         console.error(error);
-        return internalErrorResponse({ success: false, data: "Unable to remove that member" });
+        return internalErrorResponse({ success: false, data: "Unable to remove that member." });
     }
     return okResponse({ success: true, data: "Successfully removed that member" });
 }
